\name{convertHangulStringToKeyStrokes}
\alias{convertHangulStringToKeyStrokes}
\title{Convert Hangul string to keystrokes}
\usage{convertHangulStringToKeyStrokes(hangul, isFullwidth = TRUE)}
\arguments{
  \item{hangul}{Hangul sentence}
  \item{isFullwidth}{specify returned character will be Fullwidth ASCII or Halfwidth ASCII}
}
\value{Keystroke sequence}
\description{Convert Hangul syllables to two-beolsik keystroke sequences.}
