\docType{data}
\name{editweights}
\alias{editweights}
\title{Keystroke misspell cost table}
\format{A 66 by 66 numeric matrix of edit (substitution) costs indexed by keystroke characters on both dimensions.}
\source{Derived from the edit weight table distributed with 'KoNLP'.}
\value{A 66 by 66 numeric matrix of keystroke edit costs.}
\description{KoNLP-compatible edit weight metadata. HanNLP does not use this table in runtime algorithms yet.}
\keyword{datasets}
