\name{statDic}
\alias{statDic}
\title{Summary of dictionaries}
\usage{statDic(which = "current", n = 6)}
\arguments{
  \item{which}{one of \code{current} or \code{backup}}
  \item{n}{number of rows for head and tail}
}
\value{list containing summary, head, and tail}
\description{Show summary, head, and tail of current or backup user dictionaries.}
