\docType{data}
\name{tags}
\alias{tags}
\title{Tag names}
\format{A named character vector mapping each supported tag name to itself, used to validate tag entries.}
\source{Derived from the KAIST/HanNanum tag set distributed with 'KoNLP'.}
\value{A named character vector of supported tag names.}
\description{KoNLP-compatible KAIST/HanNanum tag names used by HanNLP dictionary utilities.}
\keyword{datasets}
