\name{useSystemDic}
\alias{useSystemDic}
\title{Use system default dictionary}
\usage{useSystemDic(backup = TRUE)}
\arguments{\item{backup}{backup current dictionary before switching}}
\value{invisible TRUE}
\description{Restore the writable user dictionary to HanNLP's bundled default user dictionary.}
